requirements.txt
aniso8601==10.0.1
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.2
Flask-Mail==0.10.0
Flask-RESTful==0.3.10
gunicorn==25.0.3
h11==0.16.0
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
priority==2.0.0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
resend==2.21.0
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
wsproto==1.3.2
